Don't skip parsing branches if no destination set
If no destination is set we still want to send the pings, as per the
previous comment. However, if we have branch specific pings configured,
we still need to parse the output to figure out which branches have been
updated.

So instead of aborting the parsing, move the check for no-destination to
the actual mail sending, and just skip sending the mail in this case (it
is cheap enough to generate that we don't have to care about generating
it and then throwing it away)