Allow overriding sender address of outgoing commit messages
Previously we'd always pick the email address from the committer record
except when there was none available (like branch creation) in which
case we used a fallback address.

With this we allow the specification of an override address, typically
noreply@domain.com, that will be used as the sender, but keeps the name
of the author. This way, the email can be properly DKIM signed.