Don't send commit msgs for rebased patches on excluded branches
When an excluded branch is rebased on one that is not, the same commit
shows up for multiple branches when received. Previously we sent the
message in this case, which seems wrong. Instead, if any excluded
branch shows up as any of the branches for a specific commit, consider
it excluded.